chore(common/base/range_map): move unit test to embedded mod
