+++ src/common/base/src/rangemap/range_map_key.rs
@@ -74,3 +74,67 @@ where
         Some(self.cmp(other))
     }
 }
+
+#[cfg(test)]
+mod tests {
+
+    use core::cmp::Ordering;
+    use std::collections::BTreeMap;
+    use std::collections::BTreeSet;
+
+    use crate::rangemap::RangeMapKey;
+
+    fn upsert_cmp_map(map: &mut BTreeMap<String, BTreeSet<String>>, k: String, v: String) {
+        if map.get(&k).is_none() {
+            map.insert(k.clone(), BTreeSet::new());
+        }
+        if let Some(set) = map.get_mut(&k) {
+            set.insert(v);
+        }
+    }
+
+    /// test if or not RangeKey satisfy reflexive property
+    #[test]
+    fn test_range_wrapper_reflexive_property() {
+        let mut tests = vec![];
+        for i in 0..10 {
+            tests.push(RangeMapKey::new(0..i, ()));
+        }
+
+        let mut less_map: BTreeMap<String, BTreeSet<String>> = BTreeMap::new();
+        let mut greater_map: BTreeMap<String, BTreeSet<String>> = BTreeMap::new();
+
+        // test antisymmetric propery and construct {less|greater}_map
+        // antisymmetric propery: if a > b then b > a.
+        for i in tests.iter() {
+            for j in tests.clone().iter() {
+                let ret_i_j = i.cmp(j);
+                let ret_j_i = j.cmp(i);
+
+                match ret_i_j {
+                    Ordering::Equal => {
+                        assert_eq!(ret_j_i, Ordering::Equal);
+                    }
+                    Ordering::Less => {
+                        assert_eq!(ret_j_i, Ordering::Greater);
+                        upsert_cmp_map(&mut less_map, i.to_string(), j.to_string());
+                    }
+                    Ordering::Greater => {
+                        assert_eq!(ret_j_i, Ordering::Less);
+                        upsert_cmp_map(&mut greater_map, i.to_string(), j.to_string());
+                    }
+                }
+            }
+        }
+
+        // prove transitive property: if a<b and b<c, then a<c
+        for (k, v) in less_map.iter() {
+            for g in v.iter() {
+                assert!(greater_map.get(g).is_some());
+                if let Some(set) = greater_map.get_mut(g) {
+                    assert!(set.get(k).is_some());
+                }
+            }
+        }
+    }
+}
